Resume-Parser-main/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
spacy==3.8.7
pdfplumber==0.10.3
pdfminer.six==20221105
docx2txt==0.8
pandas==2.1.3
openpyxl==3.1.2
scikit-learn==1.3.2
sentence-transformers==3.4.1
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
google-generativeai==0.3.0
python-dotenv==1.0.0
transformers==4.48.0
accelerate==1.3.0
codecarbon==3.2.3
python-multipart==0.0.12
